add department

+++ ack-sys-cms-web/src/main/java/org/ack/sys/cms/persist/mapper/MenuMapper.java
@@ -5,6 +5,6 @@ import org.ack.sys.cms.pojo.Menu;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
-public interface MenuMapper extends PageDao<Menu, Integer>{
+public interface MenuMapper extends PageDao<Menu, Long>{
 
 }
