+++ ack-sys-cms-web/src/main/java/org/ack/sys/cms/service/MenuService.java
@@ -9,7 +9,7 @@ import org.ack.sys.cms.pojo.Menu;
  * @author ack
  *
  */
-public interface MenuService extends PageService<Menu, Integer>{
+public interface MenuService extends PageService<Menu, Long>{
 
 	/**根據用户登陆名称查询用户信息
 	 * @param Menuname
